+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# pgEdge Platform Release Notes #############
+# pgEdge Labs Release Notes #############
 
 ## done for 24.1.7 as of 2024-01-30 ###########
   - bump pgvector from 0.5.1 to [0.6.0](https://github.com/pgvector/pgvector/blob/master/CHANGELOG.md)
   - improve [DB doc](https://github.com/pgEdge/cli/blob/REL24_1/cli/DB-README.md)
   - improve [ACE doc](https://github.com/pgEdge/cli/blob/REL24_1/cli/ACE-README.md) 
+  - cleanout nclibs, ddlx, & multicorn2 from versions.sql (not supported yet)